// src/components/casino/enums/casino.enum.ts
export enum CasinoID {
    EVOPLAY = 1,
    SLOTEGRATOR_BETGAMES = 3,
    SLOTEGRATOR_BETSOFT = 4,
    SLOTEGRATOR_PLAYSON = 5,
    SLOTEGRATOR_SPINOMENAL = 6,
    SLOTEGRATOR_IGROSOFT = 7,
    SLOTEGRATOR_REDRAKE = 8,
    SLOTEGRATOR_GAMEART = 9,
    SLOTEGRATOR_HABANERO = 10,
    SLOTEGRATOR_PLATIPUS = 11,
    SLOTEGRATOR_VIVOGAMING = 12,
    SLOTEGRATOR_TOMHORN = 13,
    SLOTEGRATOR_ENDORPHINA = 14,
    SLOTEGRATOR_AMATIC = 15,
    SLOTEGRATOR_BOOONGO = 16,
    SLOTEGRATOR_MICROGAMING = 17,
    SLOTEGRATOR_BIG_TIME_GAMING = 18,
    GAMSHY = 19,
    MASCOT = 20,
    PRAGMATIC = 21,
    SLOTEGRATOR_DLV = 22,
    B2BSLOTS = 23
}